Fix 'preloading' of images
The original version tried to establish a connection, but the code didn't actually work because the change to the src wasn't evaluated until the JS finished, and by that time it was too late. This delays starting the pings until all servers responded. It's onerror because the images aren't images.