Fix stop timeout
If the ping is manually stopped, it will be restarted after the 30s timeout is fired. This clears the timeout so it won't be fired until it is manually clicked again. It also keeps only one timeout active at any given time, so that you won't have waves of start/stop happening every 30 seconds if you click a bunch of times.